move test_de_densityscale
