+++ astropy/cosmology/tests/test_cosmology.py
@@ -196,72 +196,6 @@ def test_efunc_vs_invefunc_flrw():
 
 
 @pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
-def test_de_densityscale():
-    cosmo = flrw.LambdaCDM(H0=70, Om0=0.3, Ode0=0.70)
-    z = np.array([0.1, 0.2, 0.5, 1.5, 2.5])
-    assert allclose(cosmo.de_density_scale(z), [1.0, 1.0, 1.0, 1.0, 1.0])
-    # Integer check
-    assert allclose(cosmo.de_density_scale(3), cosmo.de_density_scale(3.0), rtol=1e-7)
-    assert allclose(
-        cosmo.de_density_scale([1, 2, 3]),
-        cosmo.de_density_scale([1.0, 2.0, 3.0]),
-        rtol=1e-7,
-    )
-
-    cosmo = flrw.wCDM(H0=70, Om0=0.3, Ode0=0.60, w0=-0.5)
-    assert allclose(
-        cosmo.de_density_scale(z),
-        [1.15369, 1.31453, 1.83712, 3.95285, 6.5479],
-        rtol=1e-4,
-    )
-    assert allclose(cosmo.de_density_scale(3), cosmo.de_density_scale(3.0), rtol=1e-7)
-    assert allclose(
-        cosmo.de_density_scale([1, 2, 3]),
-        cosmo.de_density_scale([1.0, 2.0, 3.0]),
-        rtol=1e-7,
-    )
-
-    cosmo = flrw.w0wzCDM(H0=70, Om0=0.3, Ode0=0.50, w0=-1, wz=0.5)
-    assert allclose(
-        cosmo.de_density_scale(z),
-        [0.746048, 0.5635595, 0.25712378, 0.026664129, 0.0035916468],
-        rtol=1e-4,
-    )
-    assert allclose(cosmo.de_density_scale(3), cosmo.de_density_scale(3.0), rtol=1e-7)
-    assert allclose(
-        cosmo.de_density_scale([1, 2, 3]),
-        cosmo.de_density_scale([1.0, 2.0, 3.0]),
-        rtol=1e-7,
-    )
-
-    cosmo = flrw.w0waCDM(H0=70, Om0=0.3, Ode0=0.70, w0=-1, wa=-0.5)
-    assert allclose(
-        cosmo.de_density_scale(z),
-        [0.9934201, 0.9767912, 0.897450, 0.622236, 0.4458753],
-        rtol=1e-4,
-    )
-    assert allclose(cosmo.de_density_scale(3), cosmo.de_density_scale(3.0), rtol=1e-7)
-    assert allclose(
-        cosmo.de_density_scale([1, 2, 3]),
-        cosmo.de_density_scale([1.0, 2.0, 3.0]),
-        rtol=1e-7,
-    )
-
-    cosmo = flrw.wpwaCDM(H0=70, Om0=0.3, Ode0=0.70, wp=-0.9, wa=0.2, zp=0.5)
-    assert allclose(
-        cosmo.de_density_scale(z),
-        [1.012246048, 1.0280102, 1.087439, 1.324988, 1.565746],
-        rtol=1e-4,
-    )
-    assert allclose(cosmo.de_density_scale(3), cosmo.de_density_scale(3.0), rtol=1e-7)
-    assert allclose(
-        cosmo.de_density_scale([1, 2, 3]),
-        cosmo.de_density_scale([1.0, 2.0, 3.0]),
-        rtol=1e-7,
-    )
-
-
-@pytest.mark.skipif(not HAS_SCIPY, reason="test requires scipy")
 def test_comoving_distance_z1z2():
     tcos = flrw.LambdaCDM(100, 0.3, 0.8, Tcmb0=0.0)
     with pytest.raises(ValueError):  # test diff size z1, z2 fail
